+++ internal/database/dbcache/cached_default_repos.go
@@ -1,4 +1,4 @@
-package backend
+package dbcache
 
 import (
 	"context"
