+++ internal/database/dbcache/cached_default_repos_test.go
@@ -1,4 +1,4 @@
-package backend
+package dbcache
 
 import (
 	"context"
